finalized changes
